+++ aerokit/aero/plot/shockpolar.py
@@ -8,7 +8,7 @@ import aerokit.aero.degree as deg
 import aerokit.aero.ShockWave as sw
 
 import matplotlib.pyplot as plt
-from aerokit.aero.plot.defaultstyle import figure_theta_pressure, figure_theta_sigma
+from aerokit.aero.plot.defaultstyle import figure_theta_pressure, figure_theta_sigma, set_grid  # noqa: F401
 
 
 def plot_theta_sigma(